+++ app/xrtd/Options.hs
@@ -7,7 +7,7 @@ import           Data.ByteArray.HexString               (HexString)
 import           Network.Ethereum.Account.LocalKey      (LocalKey (..))
 import           Network.Ethereum.Api.Provider          (Provider (..))
 import           Network.Ethereum.Chain
-import           Network.Ethereum.Unit                  (Szabo)
+import           Network.Ethereum.Unit                  (Shannon, fromWei)
 import           Options.Applicative
 
 import qualified Network.Robonomics.Lighthouse.Provider as P
@@ -33,15 +33,25 @@ providerOptions = P.Config
     <*> option str (long "lighthouse" <> value "airalab.lighthouse.5.robonomics.eth" <> metavar "ENS" <> help "Robonomics lighthouse name")
     <*> option str (long "factory" <> value "factory.5.robonomics.eth" <> metavar "ENS" <> help "Robonomics liability factory name")
     <*> option str (long "ens" <> value "0x314159265dD8dbb310642f98f50C066173C1259b" <> metavar "ADDRESS" <> help "ENS registry contract")
-    <*> option gasprice (long "gasprice" <> value (5 :: Szabo) <> metavar "GAS_PRICE" <> help "Transaction gas price, GWei [DEFAULT: 5]")
+    <*> option gasprice (long "gasprice" <> value P.SafePrice <> metavar "GAS_PRICE" <> help "Transaction gas price (safe, fast, fastest or in GWei) [DEFAULT: safe]")
 
 account :: Parser LocalKey
 account = LocalKey . (importKey :: HexString -> PrivateKey)
     <$> option str (long "private" <> metavar "KEY" <> help "Hex encoded private key")
     <*> option chain (long "chain" <> value foundation <> metavar "CHAIN_ID" <> help "Ethereum chain [foundation, ropsten, kovan, rikenby]")
 
-gasprice :: ReadM Szabo
-gasprice = fromInteger <$> auto
+gasprice :: ReadM P.GasPrice
+gasprice = eitherReader go
+  where
+    go "safe" = Right P.SafePrice
+    go "fast" = Right P.FastPrice
+    go "fastest" = Right P.FastestPrice
+    go str | valid str && (read str :: Integer) > 0 = Right . P.GasPrice . shannon $ read str
+           | otherwise = Left "GAS_PRICE shoud be [safe, fast, fastest] or in GWei"
+    valid = all $ flip elem ("1234567890" :: String)
+    shannon :: Integer -> Shannon
+    shannon  = fromWei . (* 10^9)
+
 
 chain :: ReadM Integer
 chain = eitherReader go